requirements.txt
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.17.0

